cleanig components
