+++ frontend/src/App.jsx
@@ -1,12 +1,13 @@
 
 import './App.css'
+import Login from './components/Auth/Login'
 
 function App() {
  
 
   return (
     <>
-      
+    <Login/>
     </>
   )
 }
