+++ frontend/src/components/Auth/Login.jsx
@@ -1,0 +1,88 @@
+import { useState } from 'react';
+// import { useAuth } from '../../context/AuthContext';
+// import { useNavigate } from 'react-router-dom';
+
+const Login = () => {
+  const [formData, setFormData] = useState({
+    email: '',
+    password: ''
+  });
+  const [error, setError] = useState('');
+  // const { login } = useAuth();
+  // const navigate = useNavigate();
+
+  const handleChange = (e) =>{
+    const {name, value} = e.target;
+    setFormData((prevform) =>({
+      ...prevform,
+      [name]: value
+    }))
+
+}
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    setError('');
+    
+  // const result = await login(formData.email, formData.password);
+  //   if (result.success) {
+  //     navigate('/lobby');
+  //   } else {
+  //     setError(result.error);
+  //   }
+  };
+
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-gray-50">
+      <div className="max-w-md w-full space-y-8 p-8 bg-white rounded-lg shadow">
+        <h2 className="text-center text-3xl font-bold">Signin</h2>
+        
+        {error && (
+          <div className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded">
+            {error}
+          </div>
+        )}
+
+        <form className="mt-8 space-y-6" onSubmit={handleSubmit}>
+          <div>
+            <label htmlFor="email" className="block text-sm font-medium text-gray-700">
+              Email address
+            </label>
+            <input
+              id="email"
+              name="email"
+              type="email"
+              required
+              className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md"
+              value={formData.email}
+              onChange={handleChange}
+            />
+          </div>
+
+          <div>
+            <label htmlFor="password" className="block text-sm font-medium text-gray-700">
+              Password
+            </label>
+            <input
+              id="password"
+              name="password"
+              type="password"
+              required
+              className="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md"
+              value={formData.password}
+              onChange={handleChange}
+            />
+          </div>
+
+          <button
+            type="submit"
+            className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-blue-600 hover:bg-blue-700"
+          >
+            Sign in
+          </button>
+        </form>
+      </div>
+    </div>
+  );
+};
+
+export default Login;
