--- frontend/src/context/AuthContext.js
@@ -1,67 +1,0 @@
-import { createContext, useState, useContext, useEffect } from 'react';
-
-const AuthContext = createContext({});
-
-export const AuthProvider = ({ children }) => {
-  const [user, setUser] = useState(null);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    // Check for stored token on mount
-    const token = localStorage.getItem('token');
-    if (token) {
-      fetchUserData(token);
-    } else {
-      setLoading(false);
-    }
-  }, []);
-
-  const fetchUserData = async (token) => {
-    try {
-      const response = await fetch('/api/auth/me', {
-        headers: {
-          'Authorization': `Bearer ${token}`
-        }
-      });
-      const data = await response.json();
-      setUser(data.user);
-    } catch (error) {
-      localStorage.removeItem('token');
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const login = async (email, password) => {
-    try {
-      const response = await fetch('/api/auth/login', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ email, password })
-      });
-      const data = await response.json();
-      
-      if (data.token) {
-        localStorage.setItem('token', data.token);
-        setUser(data.user);
-        return { success: true };
-      }
-      return { success: false, error: data.error };
-    } catch (error) {
-        return { success: false, error: 'Login failed' };
-    }
-  };
-
-  const logout = () => {
-    localStorage.removeItem('token');
-    setUser(null);
-  };
-
-  return (
-    <AuthContext.Provider value={{ user, login, logout, loading }}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export const useAuth = () => useContext(AuthContext);
